Mofified the production build script to include both the legacy and the modern js based on the entries and browsers
